requirements.txt
'''langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf
python-dotenv'''
'''
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-google-genai==1.0.3
sentence-transformers
faiss-cpu
pypdf
python-dotenv
'''

flask
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-google-genai==1.0.3
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv
